🎨 update compiling options on MSVC & MSVC compiling warning free.

+++ tests/fixed_bytes.cpp
@@ -155,7 +155,7 @@ TEST(fixed_bytes_le, zero)
 TEST(fixed_bytes_le, prefix_plusplus)
 {
     fixed_bytes<2, byte_numbering::lsb0> bytes;
-    for (uint16_t i: std::views::iota(0u, std::numeric_limits<uint16_t>::max()))
+    for (uint16_t i : std::views::iota(static_cast<uint16_t>(0u), std::numeric_limits<uint16_t>::max()))
     {
         ASSERT_EQ(i, bytes.to<uint16_t>());
         ++bytes;
@@ -165,7 +165,7 @@ TEST(fixed_bytes_le, prefix_plusplus)
 TEST(fixed_bytes_be, prefix_plusplus)
 {
     fixed_bytes<2, byte_numbering::msb0> bytes;
-    for (uint16_t i: std::views::iota(0u, std::numeric_limits<uint16_t>::max()))
+    for (uint16_t i : std::views::iota(static_cast<uint16_t>(0u), std::numeric_limits<uint16_t>::max()))
     {
         ASSERT_EQ(i, bytes.to<uint16_t>());
         ++bytes;
